Move testing helper functions into a more general utilities folder
Over time more utility functions will be added to the project. The
utility folder should be the default place for them to be placed
into.
